+++ cogents_tools/integrations/bu/agent/message_manager/views.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from cogents_tools.integrations.bu.llm.messages import BaseMessage
+from cogents_tools.integrations.llm import BaseMessage
 
 if TYPE_CHECKING:
     pass
