+++ cogents_tools/integrations/bu/dom/playground/multi_act.py
@@ -1,7 +1,7 @@
 from cogents_tools.integrations.bu import Agent
 from cogents_tools.integrations.bu.browser import BrowserProfile, BrowserSession
 from cogents_tools.integrations.bu.browser.types import ViewportSize
-from cogents_tools.integrations.bu.llm import ChatAzureOpenAI
+from cogents_tools.integrations.llm import ChatAzureOpenAI
 
 # Initialize the Azure OpenAI client
 llm = ChatAzureOpenAI(
