--- cogents_tools/integrations/bu/llm/google/__init__.py
@@ -1,3 +1,0 @@
-from cogents_tools.integrations.bu.llm.google.chat import ChatGoogle
-
-__all__ = ["ChatGoogle"]
