+++ cogents_tools/integrations/utils/llm_adapter.py
@@ -1,15 +1,116 @@
 import asyncio
 import logging
-from typing import Any, Dict, List, Optional
+import os
+from typing import Any, Dict, List, Optional, Union, Protocol, TypeVar, overload, runtime_checkable
+from pydantic import BaseModel
 
 logger = logging.getLogger(__name__)
 
 import dotenv
 from cogents_core.llm import BaseLLMClient, get_llm_client
 
-from cogents_tools.integrations.bu.llm.views import ChatInvokeCompletion
-
 dotenv.load_dotenv()
+
+# Define the views and messages for compatibility
+class ChatInvokeUsage(BaseModel):
+    prompt_tokens: int
+    prompt_cached_tokens: int | None = None
+    prompt_cache_creation_tokens: int | None = None
+    prompt_image_tokens: int | None = None
+    completion_tokens: int
+    total_tokens: int
+
+class ChatInvokeCompletion(BaseModel):
+    completion: Any
+    thinking: str | None = None
+    redacted_thinking: str | None = None
+    usage: ChatInvokeUsage | None = None
+
+class ModelError(Exception):
+    pass
+
+class ModelProviderError(ModelError):
+    def __init__(self, message: str, status_code: int = 502, model: str | None = None):
+        super().__init__(message, status_code)
+        self.model = model
+
+class ModelRateLimitError(ModelProviderError):
+    def __init__(self, message: str, status_code: int = 429, model: str | None = None):
+        super().__init__(message, status_code, model)
+
+# Define message types for compatibility
+from typing import Literal
+
+class ContentPartTextParam(BaseModel):
+    text: str
+    type: Literal["text"] = "text"
+
+class ContentPartRefusalParam(BaseModel):
+    refusal: str
+    type: Literal["refusal"] = "refusal"
+
+class ContentPartImageParam(BaseModel):
+    image_url: Any
+    type: Literal["image_url"] = "image_url"
+
+class _MessageBase(BaseModel):
+    role: Literal["user", "system", "assistant"]
+    cache: bool = False
+
+class UserMessage(_MessageBase):
+    role: Literal["user"] = "user"
+    content: str | list[ContentPartTextParam | ContentPartImageParam]
+    name: str | None = None
+
+class SystemMessage(_MessageBase):
+    role: Literal["system"] = "system"
+    content: str | list[ContentPartTextParam]
+    name: str | None = None
+
+class AssistantMessage(_MessageBase):
+    role: Literal["assistant"] = "assistant"
+    content: str | list[ContentPartTextParam | ContentPartRefusalParam] | None
+    name: str | None = None
+    refusal: str | None = None
+    tool_calls: list = []
+
+BaseMessage = Union[UserMessage, SystemMessage, AssistantMessage]
+ContentText = ContentPartTextParam
+ContentRefusal = ContentPartRefusalParam
+ContentImage = ContentPartImageParam
+
+# Define the BaseChatModel protocol for compatibility
+T = TypeVar("T", bound=BaseModel)
+
+@runtime_checkable
+class BaseChatModel(Protocol):
+    _verified_api_keys: bool = False
+    model: str
+
+    @property
+    def provider(self) -> str:
+        ...
+
+    @property
+    def name(self) -> str:
+        ...
+
+    @property
+    def model_name(self) -> str:
+        return self.model
+
+    @overload
+    async def ainvoke(self, messages: list[BaseMessage], output_format: None = None) -> ChatInvokeCompletion[str]:
+        ...
+
+    @overload
+    async def ainvoke(self, messages: list[BaseMessage], output_format: type[T]) -> ChatInvokeCompletion[T]:
+        ...
+
+    async def ainvoke(
+        self, messages: list[BaseMessage], output_format: type[T] | None = None
+    ) -> ChatInvokeCompletion[T] | ChatInvokeCompletion[str]:
+        ...
 
 
 def get_llm_client_browser_compatible(structured_output=True, **kwargs) -> BaseLLMClient:
@@ -262,9 +363,99 @@ class MemoryLLMAdapter(BaseLLMAdapter):
         return self.llm_client.get_embedding_dimensions()
 
 
+# Factory functions for creating chat models
+def get_llm_by_name(model_name: str) -> BaseChatModel:
+    """
+    Factory function to create LLM instances from string names with API keys from environment.
+    
+    Args:
+        model_name: String name like 'azure_gpt_4_1_mini', 'openai_gpt_4o', etc.
+    
+    Returns:
+        LLM instance with API keys from environment variables
+    
+    Raises:
+        ValueError: If model_name is not recognized
+    """
+    if not model_name:
+        raise ValueError("Model name cannot be empty")
+
+    # Parse model name
+    parts = model_name.split("_", 1)
+    if len(parts) < 2:
+        raise ValueError(f"Invalid model name format: '{model_name}'. Expected format: 'provider_model_name'")
+
+    provider = parts[0]
+    model_part = parts[1]
+
+    # Convert underscores back to dots/dashes for actual model names
+    if "gpt_4_1_mini" in model_part:
+        model = model_part.replace("gpt_4_1_mini", "gpt-4.1-mini")
+    elif "gpt_4o_mini" in model_part:
+        model = model_part.replace("gpt_4o_mini", "gpt-4o-mini")
+    elif "gpt_4o" in model_part:
+        model = model_part.replace("gpt_4o", "gpt-4o")
+    elif "gemini_2_0" in model_part:
+        model = model_part.replace("gemini_2_0", "gemini-2.0").replace("_", "-")
+    elif "gemini_2_5" in model_part:
+        model = model_part.replace("gemini_2_5", "gemini-2.5").replace("_", "-")
+    else:
+        model = model_part.replace("_", "-")
+
+    # Create the appropriate LLM client based on provider
+    if provider == "openai":
+        return BULLMAdapter(get_llm_client(provider="openai", model=model, **{"api_key": os.getenv("OPENAI_API_KEY")}))
+    elif provider == "azure":
+        return BULLMAdapter(get_llm_client(
+            provider="azure", 
+            model=model, 
+            **{
+                "api_key": os.getenv("AZURE_OPENAI_KEY") or os.getenv("AZURE_OPENAI_API_KEY"),
+                "azure_endpoint": os.getenv("AZURE_OPENAI_ENDPOINT")
+            }
+        ))
+    elif provider == "google":
+        return BULLMAdapter(get_llm_client(provider="google", model=model, **{"api_key": os.getenv("GOOGLE_API_KEY")}))
+    elif provider == "anthropic":
+        return BULLMAdapter(get_llm_client(provider="anthropic", model=model, **{"api_key": os.getenv("ANTHROPIC_API_KEY")}))
+    elif provider == "groq":
+        return BULLMAdapter(get_llm_client(provider="groq", model=model, **{"api_key": os.getenv("GROQ_API_KEY")}))
+    else:
+        available_providers = ["openai", "azure", "google", "anthropic", "groq"]
+        raise ValueError(f"Unknown provider: '{provider}'. Available providers: {', '.join(available_providers)}")
+
+
+# Create model instances on demand
+def __getattr__(name: str) -> BaseChatModel:
+    """Create model instances on demand with API keys from environment."""
+    try:
+        return get_llm_by_name(name)
+    except ValueError:
+        raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
+
+
 __all__ = [
+    # Adapters
     "BULLMAdapter",
     "MemoryLLMAdapter",
+    "BaseChatModel",
+    # Factory functions
     "get_llm_client_browser_compatible",
     "get_llm_client_memory_compatible",
+    "get_llm_by_name",
+    # Message types
+    "BaseMessage",
+    "UserMessage", 
+    "SystemMessage",
+    "AssistantMessage",
+    "ContentText",
+    "ContentRefusal", 
+    "ContentImage",
+    # Views
+    "ChatInvokeCompletion",
+    "ChatInvokeUsage",
+    # Exceptions
+    "ModelError",
+    "ModelProviderError", 
+    "ModelRateLimitError",
 ]
